BlastAndParsePAF can now work with a unique target blast db

+++ modules/Bio/EnsEMBL/Compara/RunnableDB/BlastAndParsePAF.pm
@@ -92,10 +92,19 @@ sub fetch_input {
     my $mlss            = $self->compara_dba()->get_MethodLinkSpeciesSetAdaptor->fetch_by_dbID($mlss_id) or die "Could not fetch mlss with dbID=$mlss_id";
     my $species_set     = $mlss->species_set_obj->genome_dbs;
 
-    my $genome_db_list;
+    $self->param('all_blast_db', {});
 
-    # If we restrict the search to one species at a time
-    if ($self->param('target_genome_db_id')) {
+    my $genome_db_list = [];
+
+    if ($self->param('blast_db')) {
+
+        # Target species are all mixed
+        print STDERR "Using the blast database provided: ", $self->param('blast_db'), "\n" if $self->debug;
+        $self->param('all_blast_db')->{$self->param('blast_db')} = undef;
+
+    } elsif ($self->param('target_genome_db_id')) {
+
+        # If we restrict the search to one species at a time
         $genome_db_list = [$self->compara_dba()->get_GenomeDBAdaptor->fetch_by_dbID($self->param('target_genome_db_id'))];
 
     } else {
@@ -105,7 +114,8 @@ sub fetch_input {
         my $mlss            = $self->compara_dba()->get_MethodLinkSpeciesSetAdaptor->fetch_by_dbID($mlss_id) or die "Could not fetch mlss with dbID=$mlss_id";
         my $species_set     = $mlss->species_set_obj->genome_dbs;
 
-        $genome_db_list = $species_set;
+        $genome_db_list = [ grep {$_->dbID != $self->param('genome_db_id')} @$species_set ];
+
         # If reusing this genome_db, only need to blast against the 'fresh' genome_dbs
         if ($self->param('reuse_ss_id')) {
             my $reused_species_set = $self->compara_dba()->get_SpeciesSetAdaptor->fetch_by_dbID($self->param('reuse_ss_id'));
@@ -115,13 +125,19 @@ sub fetch_input {
 #            my $reused_genome_dbs = $self->compara_dba()->get_SpeciesSetAdaptor->fetch_by_dbID($self->param('reuse_ss_id'))->genome_dbs;
             my %reuse_ss_hash = ( map { $_->dbID() => 1 } @$reused_genome_dbs );
             if ($reuse_ss_hash{$self->param('genome_db_id')}) {
-                $genome_db_list = [ grep {not $reuse_ss_hash{$_->dbID}} @$species_set ];
+                $genome_db_list = [ grep {not $reuse_ss_hash{$_->dbID}} @$genome_db_list ];
             }
         }
     }
 
     print STDERR "Found ", scalar(@$genome_db_list), " genomes to blast this member against.\n" if ($self->debug);
-    $self->param('genome_db_list', $genome_db_list);
+    my $blastdb_dir = $self->param('fasta_dir');
+    foreach my $genome_db (@$genome_db_list) {
+        my $fastafile = $blastdb_dir . '/' . $genome_db->name() . '_' . $genome_db->assembly() . '.fasta';
+        $fastafile =~ s/\s+/_/g;    # replace whitespace with '_' characters
+        $fastafile =~ s/\/\//\//g;  # converts any // in path to /
+        $self->param('all_blast_db')->{$fastafile} = $genome_db->dbID;
+    }
 
 }
 
@@ -230,45 +246,31 @@ sub run {
     $self->compara_dba->dbc->disconnect_when_inactive(1); 
 
     my $cross_pafs = [];
-    #my %cross_pafs = ();
-    foreach my $genome_db (@{$self->param('genome_db_list')}) {
-        my $fastafile = $genome_db->name() . '_' . $genome_db->assembly() . '.fasta';
-        $fastafile =~ s/\s+/_/g;    # replace whitespace with '_' characters
-            $fastafile =~ s/\/\//\//g;  # converts any // in path to /
-            my $cross_genome_dbfile = $blastdb_dir . '/' . $fastafile;   # we are always interested in the 'foreign' genome's fasta file, not the member's
+    foreach my $blast_db (keys %{$self->param('all_blast_db')}) {
+        my $target_genome_db_id = $self->param('all_blast_db')->{$blast_db};
 
-            #Don't blast against self if asked
-            if (($genome_db->dbID != $self->param('genome_db_id')) or $self->param('allow_same_species_hits')) {
-                #Run blastp
-                my $cig_cmd = $self->param('no_cigars') ? '' : 'qseq sseq';
-                my $cmd = "${blast_bin_dir}/blastp -db $cross_genome_dbfile $blast_params -evalue $evalue_limit -max_target_seqs $tophits -out $blast_outfile -outfmt '7 qacc sacc evalue score nident pident qstart qend sstart send length positive ppos $cig_cmd'";
-                if($debug) {
-                    warn "CMD:\t$cmd\n";
-                }
-                my $start_time = time();
-                open( BLAST, "| $cmd") || die qq{could not execute "$cmd", returned error code: $!};
-                $self->param('query_set')->print_sequences_to_file(-fh => \*BLAST, -format => 'fasta');
-                close BLAST;
+        my $cig_cmd = $self->param('no_cigars') ? '' : 'qseq sseq';
+        my $cmd = "$blast_bin_dir/blastp -db $blast_db $blast_params -evalue $evalue_limit -max_target_seqs $tophits -out $blast_outfile -outfmt '7 qacc sacc evalue score nident pident qstart qend sstart send length positive ppos $cig_cmd'";
+        warn "CMD:\t$cmd\n" if $debug;
 
-                print "Time for blast " . (time() - $start_time) . "\n";
+        my $start_time = time();
+        open( BLAST, "| $cmd") || die qq{could not execute "$cmd", returned error code: $!};
+        $self->param('query_set')->print_sequences_to_file(-fh => \*BLAST, -format => 'fasta');
+        close BLAST;
+        print "Time for blast " . (time() - $start_time) . "\n";
 
-                my $features = $self->parse_blast_table_into_paf($blast_outfile, $self->param('genome_db_id'), $genome_db->dbID);
-                if (defined $features) {
-                    foreach my $qmember_id (keys %$features) {
-                        my $qfeatures = $features->{$qmember_id};
-                        push @$cross_pafs, @$qfeatures;
-                        #push @{$cross_pafs{$genome_db->dbID}}, $feature;
-                    }
-                }
-                unless($debug) {
-                    unlink $blast_outfile;
-                }
+        my $features = $self->parse_blast_table_into_paf($blast_outfile, $self->param('genome_db_id'), $target_genome_db_id);
+        if (defined $features) {
+            foreach my $qmember_id (keys %$features) {
+                my $qfeatures = $features->{$qmember_id};
+                push @$cross_pafs, @$qfeatures;
             }
+        }
+        unlink $blast_outfile unless $debug;
     }
     $self->compara_dba->dbc->disconnect_when_inactive(0); 
 
     $self->param('cross_pafs', $cross_pafs);
-    #$self->param('cross_pafs', \%cross_pafs);
 }
 
 sub write_output {
